+++ packages/clerk-js/src/ui/components/UserProfile/ConnectedAccountsSection.tsx
@@ -1,4 +1,4 @@
-import type { OAuthProvider, OAuthStrategy } from '@clerk/types';
+import type { OAuthProvider, OAuthScope, OAuthStrategy } from '@clerk/types';
 import type { ExternalAccountResource } from '@clerk/types/src';
 
 import { useRouter } from '../../../ui/router';
@@ -149,7 +149,10 @@ const ConnectedAccountAccordion = ({ account }: { account: ExternalAccountResour
   );
 };
 
-function findAdditionalScopes(account: ExternalAccountResource, scopes?: Record<OAuthProvider, string[]>): string[] {
+function findAdditionalScopes(
+  account: ExternalAccountResource,
+  scopes?: Partial<Record<OAuthProvider, OAuthScope[]>>,
+): string[] {
   if (!scopes) {
     return [];
   }