+++ packages/clerk-js/src/ui/hooks/index.ts
@@ -4,7 +4,6 @@ export * from './useMagicLink';
 export * from './useClipboard';
 export * from './useEnabledThirdPartyProviders';
 export * from './useLoadingStatus';
-export * from './useSafeLayoutEffect';
 export * from './usePopover';
 export * from './useLocalStorage';
 export * from './useSearchInput';