+++ packages/clerk-js/src/ui/index.ts
@@ -1,2 +1,0 @@
-export { mountComponentRenderer } from './Components';
-export type { ComponentControls, MountComponentRenderer } from './Components';