+++ packages/expo/src/singleton.ts
@@ -1,4 +1,4 @@
-import type { FapiRequestInit, FapiResponse } from '@clerk/clerk-js/dist/types/src/core/fapiClient';
+import type { FapiRequestInit, FapiResponse } from '@clerk/clerk-js/dist/types/core/fapiClient';
 import Clerk from '@clerk/clerk-js/headless';
 import type { HeadlessBrowserClerk } from '@clerk/clerk-react';
 
@@ -17,7 +17,7 @@ export function buildClerk({ key, tokenCache }: BuildClerkOptions): HeadlessBrow
   if (!clerk) {
     const getToken = tokenCache.getToken;
     const saveToken = tokenCache.saveToken;
-
+    // TODO: DO NOT ACCEPT THIS
     clerk = new Clerk(key);
 
     // @ts-expect-error