+++ packages/shared/src/hooks/index.ts
@@ -1,5 +1,0 @@
-export * from './createContextAndHook';
-export * from './contexts';
-export * from './useOrganization';
-export * from './useOrganizationList';
-export * from './useOrganizations';